+++ iam/app/textract/locals.tf
@@ -1,0 +1,5 @@
+locals {
+  output_data = {
+    textract_analyze = module.analyze_policy.data["textract_analyze"]
+  }
+}