+++ iam/app/textract/main.tf
@@ -1,6 +1,10 @@
-module "this_table" {
-  source                          = "../../dynamodb/table"
-  monitor_data                    = var.monitor_data
+module "analyze_policy" {
+  source = "../../../iam/policy/identity/textract"
+  policy_map = {
+    textract_analyze = {
+      adapter_name = "*"
+    }
+  }
   name_append_default             = var.name_append_default
   name_include_app_fields_default = var.name_include_app_fields_default
   name_infix_default              = var.name_infix_default
@@ -8,13 +12,4 @@ module "this_table" {
   name_prepend_default            = var.name_prepend_default
   name_suffix_default             = var.name_suffix_default
   std_map                         = var.std_map
-  table_map = {
-    (local.table_name) = {
-      attribute_map = {
-        LockID = "S"
-      }
-      hash_key = "LockID"
-    }
-  }
-  table_server_side_encryption_enabled_default = var.table_server_side_encryption_enabled_default
 }